requirements.txt
smolagents
numpy
openai
python-dotenv
